# [jdbc.read] Don't track realized keys in TransientRow