Tracking the hashset of realized keys for each row in the resultset has noticeable overhead. Since this is only needed to print the transient row during debugging (as far as I can tell), it makes sense to shave the overhead here and just print the persistent version of the transient row if printing is needed. Please tell me if I'm wrong or missing something.