@alexander-yakushev 2024-10-08
Update regarding keeping tests: the current tests are written using `:venues` model that has columns `[:id :name :category :created-at :updated-at]`, but e.g. the assoc test uses `:k` as a key and then checks that it is among realized-keys. Is this testing the intentional behavior that you want to preserve?