@alexander-yakushev 2024-10-08
Here's the benchmark:

```clj
;; preparation - insert 10k rows into people table
(toucan2.insert/insert!
 ::test/people
 (for [id (range 10 10000)]
   (instance/instance ::test/people
                      {:id id
                       :name       "Somebody"
                       :created-at (LocalDateTime/parse "2017-01-01T00:00")})))

;; the benchmark - I use time+ to get allocations too, but criterium would work just as well
(time+ (select/select ::test/people {:select [:*], :from [[:people]]}))
```

Results:

```
Before:
Time per call: 18.89 ms   Alloc per call: 17,484,278b
Time per call: 18.90 ms   Alloc per call: 17,483,998b
Time per call: 18.87 ms   Alloc per call: 17,483,817b

After:
Time per call: 15.17 ms   Alloc per call: 16,771,048b
Time per call: 15.14 ms   Alloc per call: 16,770,815b
Time per call: 15.18 ms   Alloc per call: 16,770,697b
```