@alexander-yakushev 2024-10-08
And another follow-up on having to calculate all columns ahead of time: I'm mostly looking at Metabase benchmarks (not microbenchmarks of Toucan2 itself) and I see that in Metabase `fetch-all-columns!` is usually called; hence, all column names are eventually calculated anyway.