@alexander-yakushev 2025-01-05
### Update

Here's a completely different approach. Don't track realized keys, don't pre-compute column names. Use a bit of reflection to work around the inability to dig into the keys of a transient collection. Dig into the private underlying array of a transient map, get the values from there. Not entirely safe indeed, but this is a devtime thing, and it's already protected with a try/catch.

Benchmark:
```
- master:
Time per call: 3.91 ms   Alloc per call: 8,448,707b   Iterations: 2577
Time per call: 3.84 ms   Alloc per call: 8,447,663b   Iterations: 2647
Time per call: 3.93 ms   Alloc per call: 8,447,364b   Iterations: 2703

- no-realized-keys:
Time per call: 3.60 ms   Alloc per call: 7,888,220b   Iterations: 2776
Time per call: 3.60 ms   Alloc per call: 7,887,431b   Iterations: 2781
Time per call: 3.70 ms   Alloc per call: 7,886,683b   Iterations: 2702
```

-5% time, -7% allocations, and simplifying the main path of the implementation which is the biggest win.